pyproject.toml
[tool.poetry]
name = "overwatch-index"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^1.12.8"
google-auth-oauthlib = "^0.4.2"
google-auth-httplib2 = "^0.0.4"
Jinja2 = "^2.11.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = {version = "^20.8b1", allow-prereleases = true}
ghp-import = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
